pyproject.toml
[tool.poetry]
name = "repl_nix_py-2"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
django-import-export = "^3.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
